+++ packages/bridge-ui/src/proof/ProofService.ts
@@ -1,7 +1,7 @@
 import { Contract, ethers } from 'ethers';
 import { RLP } from 'ethers/lib/utils.js';
 import { crossChainSyncABI } from '../constants/abi';
-import type { Block, BlockHeader } from '../domain/block';
+import type { Block } from '../domain/block';
 import type {
   Prover,
   GenerateProofOpts,
@@ -35,53 +35,35 @@ export class ProofService implements Prover {
     return key;
   }
 
-  private static async getBlockAndBlockHeader(
+  private static async getBlock(
     contract: ethers.Contract,
     provider: ethers.providers.StaticJsonRpcProvider,
-  ): Promise<{ block: Block; blockHeader: BlockHeader }> {
-    const latestSyncedHeader = await contract.getCrossChainBlockHash(0);
+  ): Promise<Block> {
+    const latestBlockHash = await contract.getCrossChainBlockHash(0);
 
     const block: Block = await provider.send('eth_getBlockByHash', [
-      latestSyncedHeader,
+      latestBlockHash,
       false,
     ]);
 
-    const logsBloom = block.logsBloom.toString().substring(2);
-
-    const blockHeader: BlockHeader = {
-      parentHash: block.parentHash,
-      ommersHash: block.sha3Uncles,
-      beneficiary: block.miner,
-      stateRoot: block.stateRoot,
-      transactionsRoot: block.transactionsRoot,
-      receiptsRoot: block.receiptsRoot,
-      logsBloom: logsBloom.match(/.{1,64}/g)!.map((s: string) => '0x' + s),
-      difficulty: block.difficulty,
-      height: block.number,
-      gasLimit: block.gasLimit,
-      gasUsed: block.gasUsed,
-      timestamp: block.timestamp,
-      extraData: block.extraData,
-      mixHash: block.mixHash,
-      nonce: block.nonce,
-      baseFeePerGas: block.baseFeePerGas ? parseInt(block.baseFeePerGas) : 0,
-      withdrawalsRoot: block.withdrawalsRoot ?? ethers.constants.HashZero,
-    };
-
-    return { block, blockHeader };
+    return block;
   }
 
   private static getSignalProof(
     proof: EthGetProofResponse,
-    blockHeader: BlockHeader,
+    blockHeight: number,
   ) {
     // RLP encode the proof together for LibTrieProof to decode
     const encodedProof = RLP.encode(proof.storageProof[0].proof);
 
-    // encode the SignalProof struct from LibBridgeSignal
+    // Encode the SignalProof struct:
+    // struct SignalProof {
+    //   uint256 height;
+    //   bytes proof;
+    // }
     const signalProof = ethers.utils.defaultAbiCoder.encode(
       ['tuple(uint256 height, bytes proof)'],
-      [{ height: blockHeader.height, proof: encodedProof }],
+      [{ height: blockHeight, proof: encodedProof }],
     );
 
     return signalProof;
@@ -98,10 +80,7 @@ export class ProofService implements Prover {
       this.providers[opts.destChain],
     );
 
-    const { block, blockHeader } = await ProofService.getBlockAndBlockHeader(
-      contract,
-      provider,
-    );
+    const block = await ProofService.getBlock(contract, provider);
 
     // rpc call to get the merkle proof what value is at key on the SignalService contract
     const proof: EthGetProofResponse = await provider.send('eth_getProof', [
@@ -116,7 +95,7 @@ export class ProofService implements Prover {
       throw Error('invalid proof');
     }
 
-    const signalProof = ProofService.getSignalProof(proof, blockHeader);
+    const signalProof = ProofService.getSignalProof(proof, block.number);
 
     return signalProof;
   }
@@ -132,10 +111,7 @@ export class ProofService implements Prover {
       this.providers[opts.srcChain],
     );
 
-    const { block, blockHeader } = await ProofService.getBlockAndBlockHeader(
-      contract,
-      provider,
-    );
+    const block = await ProofService.getBlock(contract, provider);
 
     // rpc call to get the merkle proof what value is at key on the SignalService contract
     const proof: EthGetProofResponse = await provider.send('eth_getProof', [
@@ -150,7 +126,7 @@ export class ProofService implements Prover {
       throw Error('invalid proof');
     }
 
-    const signalProof = ProofService.getSignalProof(proof, blockHeader);
+    const signalProof = ProofService.getSignalProof(proof, block.number);
 
     return signalProof;
   }