+++ packages/protocol/contracts/bridge/BridgedERC20.sol
@@ -10,10 +10,8 @@ import {
     IERC20Upgradeable,
     ERC20Upgradeable
 } from "@openzeppelin/contracts-upgradeable/token/ERC20/ERC20Upgradeable.sol";
-// solhint-disable-next-line max-line-length
-import {
-    IERC20MetadataUpgradeable
-} from "@openzeppelin/contracts-upgradeable/token/ERC20/extensions/IERC20MetadataUpgradeable.sol";
+import {IERC20MetadataUpgradeable} from
+    "@openzeppelin/contracts-upgradeable/token/ERC20/extensions/IERC20MetadataUpgradeable.sol";
 
 import {EssentialContract} from "../common/EssentialContract.sol";
 import {Proxied} from "../common/Proxied.sol";
@@ -27,13 +25,25 @@ contract BridgedERC20 is
     ERC20Upgradeable,
     BridgeErrors
 {
+    /*//////////////////////////////////////////////////////////////
+                            STATE VARIABLES
+    //////////////////////////////////////////////////////////////*/
+
     address public srcToken;
     uint256 public srcChainId;
     uint8 private srcDecimals;
     uint256[47] private __gap;
 
+    /*//////////////////////////////////////////////////////////////
+                                 EVENTS
+    //////////////////////////////////////////////////////////////*/
+
     event BridgeMint(address indexed account, uint256 amount);
     event BridgeBurn(address indexed account, uint256 amount);
+
+    /*//////////////////////////////////////////////////////////////
+                         USER-FACING FUNCTIONS
+    //////////////////////////////////////////////////////////////*/
 
     /// @dev Initializer to be called after being deployed behind a proxy.
     // Intention is for a different BridgedERC20 Contract to be deployed
@@ -47,11 +57,8 @@ contract BridgedERC20 is
         string memory _name
     ) external initializer {
         if (
-            _srcToken == address(0) ||
-            _srcChainId == 0 ||
-            _srcChainId == block.chainid ||
-            bytes(_symbol).length == 0 ||
-            bytes(_name).length == 0
+            _srcToken == address(0) || _srcChainId == 0 || _srcChainId == block.chainid
+                || bytes(_symbol).length == 0 || bytes(_name).length == 0
         ) {
             revert B_INIT_PARAM_ERROR();
         }
@@ -63,29 +70,24 @@ contract BridgedERC20 is
     }
 
     /// @dev only a TokenVault can call this function
-    function bridgeMintTo(
-        address account,
-        uint256 amount
-    ) public onlyFromNamed("token_vault") {
+    function bridgeMintTo(address account, uint256 amount) public onlyFromNamed("token_vault") {
         _mint(account, amount);
         emit BridgeMint(account, amount);
     }
 
     /// @dev only a TokenVault can call this function
-    function bridgeBurnFrom(
-        address account,
-        uint256 amount
-    ) public onlyFromNamed("token_vault") {
+    function bridgeBurnFrom(address account, uint256 amount) public onlyFromNamed("token_vault") {
         _burn(account, amount);
         emit BridgeBurn(account, amount);
     }
 
     /// @dev any address can call this
     // caller must have at least amount to call this
-    function transfer(
-        address to,
-        uint256 amount
-    ) public override(ERC20Upgradeable, IERC20Upgradeable) returns (bool) {
+    function transfer(address to, uint256 amount)
+        public
+        override(ERC20Upgradeable, IERC20Upgradeable)
+        returns (bool)
+    {
         if (to == address(this)) {
             revert B_ERC20_CANNOT_RECEIVE();
         }
@@ -95,11 +97,11 @@ contract BridgedERC20 is
     /// @dev any address can call this
     // caller must have allowance of at least 'amount'
     // for 'from's tokens.
-    function transferFrom(
-        address from,
-        address to,
-        uint256 amount
-    ) public override(ERC20Upgradeable, IERC20Upgradeable) returns (bool) {
+    function transferFrom(address from, address to, uint256 amount)
+        public
+        override(ERC20Upgradeable, IERC20Upgradeable)
+        returns (bool)
+    {
         if (to == address(this)) {
             revert B_ERC20_CANNOT_RECEIVE();
         }