+++ packages/protocol/contracts/common/IMintableERC20.sol
@@ -6,9 +6,8 @@
 
 pragma solidity ^0.8.18;
 
-import {
-    IERC20Upgradeable
-} from "@openzeppelin/contracts-upgradeable/token/ERC20/IERC20Upgradeable.sol";
+import {IERC20Upgradeable} from
+    "@openzeppelin/contracts-upgradeable/token/ERC20/IERC20Upgradeable.sol";
 
 interface IMintableERC20 is IERC20Upgradeable {
     function mint(address account, uint256 amount) external;