+++ packages/protocol/test/TaikoL1LibTokenomicsTestnet.t.sol
@@ -15,12 +15,7 @@ import {TaikoL1TestBase} from "./TaikoL1TestBase.t.sol";
 import {LibLn} from "./LibLn.sol";
 
 contract TaikoL1WithTestnetConfig is TaikoL1 {
-    function getConfig()
-        public
-        pure
-        override
-        returns (TaikoData.Config memory config)
-    {
+    function getConfig() public pure override returns (TaikoData.Config memory config) {
         config = TaikoConfig.getConfig();
 
         config.txListCacheExpiry = 5 minutes;
@@ -42,17 +37,14 @@ contract TaikoL1LibTokenomicsTestnet is TaikoL1TestBase {
         uint16 proofTimeTarget = 120; // Approx. testnet value
         // Calculating it for our needs based on testnet/mainnet proof vars.
         // See Brecht's comment https://github.com/taikoxyz/taiko-mono/pull/13564
-        initProofTimeIssued = LibLn.calcInitProofTimeIssued(
-            feeBase,
-            proofTimeTarget,
-            ADJUSTMENT_QUOTIENT
-        );
+        initProofTimeIssued =
+            LibLn.calcInitProofTimeIssued(feeBase, proofTimeTarget, ADJUSTMENT_QUOTIENT);
 
         TaikoL1TestBase.setUp();
 
-        depositTaikoToken(Alice, 1E6 * 1E8, 100 ether);
-        depositTaikoToken(Bob, 1E6 * 1E8, 100 ether);
-        depositTaikoToken(Carol, 1E6 * 1E8, 100 ether);
+        depositTaikoToken(Alice, 1e6 * 1e8, 100 ether);
+        depositTaikoToken(Bob, 1e6 * 1e8, 100 ether);
+        depositTaikoToken(Carol, 1e6 * 1e8, 100 ether);
     }
 
     /// @dev Test what happens when proof time increases
@@ -61,9 +53,9 @@ contract TaikoL1LibTokenomicsTestnet is TaikoL1TestBase {
     {
         mine(1);
 
-        depositTaikoToken(Alice, 1E6 * 1E8, 100 ether);
-        depositTaikoToken(Bob, 1E6 * 1E8, 100 ether);
-        depositTaikoToken(Carol, 1E6 * 1E8, 100 ether);
+        depositTaikoToken(Alice, 1e6 * 1e8, 100 ether);
+        depositTaikoToken(Bob, 1e6 * 1e8, 100 ether);
+        depositTaikoToken(Carol, 1e6 * 1e8, 100 ether);
 
         bytes32 parentHash = GENESIS_BLOCK_HASH;
 
@@ -75,16 +67,12 @@ contract TaikoL1LibTokenomicsTestnet is TaikoL1TestBase {
         console2.log("Bob balance:", Bob_start_balance);
 
         for (uint256 blockId = 1; blockId < 10; blockId++) {
-            TaikoData.BlockMetadata memory meta = proposeBlock(
-                Alice,
-                1000000,
-                1024
-            );
+            TaikoData.BlockMetadata memory meta = proposeBlock(Alice, 1000000, 1024);
             uint64 proposedAt = uint64(block.timestamp);
             mine(blockId);
 
-            bytes32 blockHash = bytes32(1E10 + blockId);
-            bytes32 signalRoot = bytes32(1E9 + blockId);
+            bytes32 blockHash = bytes32(1e10 + blockId);
+            bytes32 signalRoot = bytes32(1e9 + blockId);
             proveBlock(
                 Bob,
                 Bob,
@@ -96,10 +84,7 @@ contract TaikoL1LibTokenomicsTestnet is TaikoL1TestBase {
                 signalRoot
             );
             uint64 provenAt = uint64(block.timestamp);
-            console2.log(
-                "Proof reward is:",
-                L1.getProofReward(provenAt, proposedAt)
-            );
+            console2.log("Proof reward is:", L1.getProofReward(provenAt, proposedAt));
 
             verifyBlock(Carol, 1);
             // This is where new fee evaluated
@@ -120,32 +105,16 @@ contract TaikoL1LibTokenomicsTestnet is TaikoL1TestBase {
 
         // Run another session with huge times
         for (uint256 blockId = 1; blockId < 10; blockId++) {
-            TaikoData.BlockMetadata memory meta = proposeBlock(
-                Alice,
-                1000000,
-                1024
-            );
+            TaikoData.BlockMetadata memory meta = proposeBlock(Alice, 1000000, 1024);
             uint64 proposedAt = uint64(block.timestamp);
             mine_huge();
 
-            bytes32 blockHash = bytes32(1E10 + blockId);
-            bytes32 signalRoot = bytes32(1E9 + blockId);
+            bytes32 blockHash = bytes32(1e10 + blockId);
+            bytes32 signalRoot = bytes32(1e9 + blockId);
 
-            proveBlock(
-                Bob,
-                Bob,
-                meta,
-                parentHash,
-                1000000,
-                1000000,
-                blockHash,
-                signalRoot
-            );
+            proveBlock(Bob, Bob, meta, parentHash, 1000000, 1000000, blockHash, signalRoot);
             uint64 provenAt = uint64(block.timestamp);
-            console2.log(
-                "Proof reward is:",
-                L1.getProofReward(provenAt, proposedAt)
-            );
+            console2.log("Proof reward is:", L1.getProofReward(provenAt, proposedAt));
             verifyBlock(Carol, 1);
             // This is where new fee evaluated
             printVariables("after verify");
@@ -161,14 +130,12 @@ contract TaikoL1LibTokenomicsTestnet is TaikoL1TestBase {
     }
 
     /// @dev Test what happens when proof time hectic couple of proposes, without prove, then some proofs
-    function test_balanced_state_reward_and_fee_if_proof_time_hectic()
-        external
-    {
+    function test_balanced_state_reward_and_fee_if_proof_time_hectic() external {
         mine(1);
         //Needs lot of token here - because there is lots of time elapsed between 2 'propose' blocks, which will raise the fee
-        depositTaikoToken(Alice, 1E8 * 1E8, 100 ether);
-        depositTaikoToken(Bob, 1E8 * 1E8, 100 ether);
-        depositTaikoToken(Carol, 1E8 * 1E8, 100 ether);
+        depositTaikoToken(Alice, 1e8 * 1e8, 100 ether);
+        depositTaikoToken(Bob, 1e8 * 1e8, 100 ether);
+        depositTaikoToken(Carol, 1e8 * 1e8, 100 ether);
 
         TaikoData.BlockMetadata[] memory metas = new TaikoData.BlockMetadata[](
             20
@@ -198,8 +165,8 @@ contract TaikoL1LibTokenomicsTestnet is TaikoL1TestBase {
         mine(6);
         // Prove and verify
         for (uint256 blockId = 1; blockId < 10; blockId++) {
-            bytes32 blockHash = bytes32(1E10 + blockId);
-            bytes32 signalRoot = bytes32(1E9 + blockId);
+            bytes32 blockHash = bytes32(1e10 + blockId);
+            bytes32 signalRoot = bytes32(1e9 + blockId);
             proveBlock(
                 Bob,
                 Bob,
@@ -211,10 +178,7 @@ contract TaikoL1LibTokenomicsTestnet is TaikoL1TestBase {
                 signalRoot
             );
             uint64 provenAt = uint64(block.timestamp);
-            console2.log(
-                "Proof reward is:",
-                L1.getProofReward(provenAt, proposedAtArr[blockId])
-            );
+            console2.log("Proof reward is:", L1.getProofReward(provenAt, proposedAtArr[blockId]));
 
             verifyBlock(Carol, 1);
 
@@ -231,32 +195,16 @@ contract TaikoL1LibTokenomicsTestnet is TaikoL1TestBase {
         // Run another iteration
         for (uint256 blockId = 1; blockId < 10; blockId++) {
             printVariables("before propose");
-            TaikoData.BlockMetadata memory meta = proposeBlock(
-                Alice,
-                1000000,
-                1024
-            );
+            TaikoData.BlockMetadata memory meta = proposeBlock(Alice, 1000000, 1024);
             uint64 proposedAt = uint64(block.timestamp);
             printVariables("after propose");
             mine_proofTime();
 
-            bytes32 blockHash = bytes32(1E10 + blockId);
-            bytes32 signalRoot = bytes32(1E9 + blockId);
-            proveBlock(
-                Bob,
-                Bob,
-                meta,
-                parentHash,
-                1000000,
-                1000000,
-                blockHash,
-                signalRoot
-            );
+            bytes32 blockHash = bytes32(1e10 + blockId);
+            bytes32 signalRoot = bytes32(1e9 + blockId);
+            proveBlock(Bob, Bob, meta, parentHash, 1000000, 1000000, blockHash, signalRoot);
             uint64 provenAt = uint64(block.timestamp);
-            console2.log(
-                "Proof reward is:",
-                L1.getProofReward(provenAt, proposedAt)
-            );
+            console2.log("Proof reward is:", L1.getProofReward(provenAt, proposedAt));
             verifyBlock(Carol, 1);
 
             parentHash = blockHash;
@@ -273,14 +221,13 @@ contract TaikoL1LibTokenomicsTestnet is TaikoL1TestBase {
     }
 
     /// @dev Test and see what happens when proof time is stable below the target and proving consecutive
-    function test_balanced_state_reward_and_fee_if_proof_time_stable_below_target_prooving_consecutive()
-        external
-    {
+    function test_balanced_state_reward_and_fee_if_proof_time_stable_below_target_prooving_consecutive(
+    ) external {
         mine(1);
 
-        depositTaikoToken(Alice, 1E6 * 1E8, 100 ether);
-        depositTaikoToken(Bob, 1E6 * 1E8, 100 ether);
-        depositTaikoToken(Carol, 1E6 * 1E8, 100 ether);
+        depositTaikoToken(Alice, 1e6 * 1e8, 100 ether);
+        depositTaikoToken(Bob, 1e6 * 1e8, 100 ether);
+        depositTaikoToken(Carol, 1e6 * 1e8, 100 ether);
 
         bytes32 parentHash = GENESIS_BLOCK_HASH;
 
@@ -294,17 +241,13 @@ contract TaikoL1LibTokenomicsTestnet is TaikoL1TestBase {
         //parentHash = prove_with_increasing_time(parentHash, 10);
         for (uint256 blockId = 1; blockId < 10; blockId++) {
             printVariables("before propose");
-            TaikoData.BlockMetadata memory meta = proposeBlock(
-                Alice,
-                1000000,
-                1024
-            );
+            TaikoData.BlockMetadata memory meta = proposeBlock(Alice, 1000000, 1024);
             uint64 proposedAt = uint64(block.timestamp);
             printVariables("after propose");
             mine(2);
 
-            bytes32 blockHash = bytes32(1E10 + blockId);
-            bytes32 signalRoot = bytes32(1E9 + blockId);
+            bytes32 blockHash = bytes32(1e10 + blockId);
+            bytes32 signalRoot = bytes32(1e9 + blockId);
             proveBlock(
                 Bob,
                 Bob,
@@ -316,10 +259,7 @@ contract TaikoL1LibTokenomicsTestnet is TaikoL1TestBase {
                 signalRoot
             );
             uint64 provenAt = uint64(block.timestamp);
-            console2.log(
-                "Proof reward is:",
-                L1.getProofReward(provenAt, proposedAt)
-            );
+            console2.log("Proof reward is:", L1.getProofReward(provenAt, proposedAt));
 
             verifyBlock(Carol, 1);
 
@@ -338,14 +278,13 @@ contract TaikoL1LibTokenomicsTestnet is TaikoL1TestBase {
     }
 
     /// @dev Test and see what happens when proof time is stable below the target and proving non consecutive
-    function test_balanced_state_reward_and_fee_if_proof_time_stable_below_target_proving_non_consecutive()
-        external
-    {
+    function test_balanced_state_reward_and_fee_if_proof_time_stable_below_target_proving_non_consecutive(
+    ) external {
         mine(1);
 
-        depositTaikoToken(Alice, 1E6 * 1E8, 100 ether);
-        depositTaikoToken(Bob, 1E6 * 1E8, 100 ether);
-        depositTaikoToken(Carol, 1E6 * 1E8, 100 ether);
+        depositTaikoToken(Alice, 1e6 * 1e8, 100 ether);
+        depositTaikoToken(Bob, 1e6 * 1e8, 100 ether);
+        depositTaikoToken(Carol, 1e6 * 1e8, 100 ether);
 
         TaikoData.BlockMetadata[] memory meta = new TaikoData.BlockMetadata[](
             30
@@ -373,8 +312,8 @@ contract TaikoL1LibTokenomicsTestnet is TaikoL1TestBase {
         mine(6);
         //Prove and verify
         for (uint256 blockId = 1; blockId < 30; blockId++) {
-            bytes32 blockHash = bytes32(1E10 + blockId);
-            bytes32 signalRoot = bytes32(1E9 + blockId);
+            bytes32 blockHash = bytes32(1e10 + blockId);
+            bytes32 signalRoot = bytes32(1e9 + blockId);
 
             proveBlock(
                 Bob,
@@ -388,10 +327,7 @@ contract TaikoL1LibTokenomicsTestnet is TaikoL1TestBase {
             );
 
             uint64 provenAt = uint64(block.timestamp);
-            console2.log(
-                "Proof reward is:",
-                L1.getProofReward(provenAt, proposedAt[blockId])
-            );
+            console2.log("Proof reward is:", L1.getProofReward(provenAt, proposedAt[blockId]));
 
             verifyBlock(Carol, 1);
 
@@ -411,14 +347,12 @@ contract TaikoL1LibTokenomicsTestnet is TaikoL1TestBase {
     }
 
     /// @dev Test what happens when proof time decreases
-    function test_balanced_state_reward_and_fee_if_proof_time_decreases()
-        external
-    {
+    function test_balanced_state_reward_and_fee_if_proof_time_decreases() external {
         mine(1);
 
-        depositTaikoToken(Alice, 1E6 * 1E8, 100 ether);
-        depositTaikoToken(Bob, 1E6 * 1E8, 100 ether);
-        depositTaikoToken(Carol, 1E6 * 1E8, 100 ether);
+        depositTaikoToken(Alice, 1e6 * 1e8, 100 ether);
+        depositTaikoToken(Bob, 1e6 * 1e8, 100 ether);
+        depositTaikoToken(Carol, 1e6 * 1e8, 100 ether);
 
         TaikoData.BlockMetadata[] memory meta = new TaikoData.BlockMetadata[](
             20
@@ -447,8 +381,8 @@ contract TaikoL1LibTokenomicsTestnet is TaikoL1TestBase {
         mine(6);
         // Prove and verify
         for (uint256 blockId = 1; blockId < 20; blockId++) {
-            bytes32 blockHash = bytes32(1E10 + blockId);
-            bytes32 signalRoot = bytes32(1E9 + blockId);
+            bytes32 blockHash = bytes32(1e10 + blockId);
+            bytes32 signalRoot = bytes32(1e9 + blockId);
             proveBlock(
                 Bob,
                 Bob,
@@ -461,10 +395,7 @@ contract TaikoL1LibTokenomicsTestnet is TaikoL1TestBase {
             );
 
             uint64 provenAt = uint64(block.timestamp);
-            console2.log(
-                "Proof reward is:",
-                L1.getProofReward(provenAt, proposedAt[blockId])
-            );
+            console2.log("Proof reward is:", L1.getProofReward(provenAt, proposedAt[blockId]));
 
             verifyBlock(Carol, 1);
             mine(21 - blockId);
@@ -481,14 +412,13 @@ contract TaikoL1LibTokenomicsTestnet is TaikoL1TestBase {
     }
 
     /// @dev Test and see what happens when proof time is stable above the target and proving consecutive
-    function test_balanced_state_reward_and_fee_if_proof_time_stable_above_target_prooving_consecutive()
-        external
-    {
+    function test_balanced_state_reward_and_fee_if_proof_time_stable_above_target_prooving_consecutive(
+    ) external {
         mine(1);
 
-        depositTaikoToken(Alice, 1E6 * 1E8, 100 ether);
-        depositTaikoToken(Bob, 1E6 * 1E8, 100 ether);
-        depositTaikoToken(Carol, 1E6 * 1E8, 100 ether);
+        depositTaikoToken(Alice, 1e6 * 1e8, 100 ether);
+        depositTaikoToken(Bob, 1e6 * 1e8, 100 ether);
+        depositTaikoToken(Carol, 1e6 * 1e8, 100 ether);
 
         bytes32 parentHash = GENESIS_BLOCK_HASH;
 
@@ -502,17 +432,13 @@ contract TaikoL1LibTokenomicsTestnet is TaikoL1TestBase {
         //parentHash = prove_with_increasing_time(parentHash, 10);
         for (uint256 blockId = 1; blockId < 10; blockId++) {
             printVariables("before propose");
-            TaikoData.BlockMetadata memory meta = proposeBlock(
-                Alice,
-                1000000,
-                1024
-            );
+            TaikoData.BlockMetadata memory meta = proposeBlock(Alice, 1000000, 1024);
             uint64 proposedAt = uint64(block.timestamp);
             printVariables("after propose");
             mine(5);
 
-            bytes32 blockHash = bytes32(1E10 + blockId);
-            bytes32 signalRoot = bytes32(1E9 + blockId);
+            bytes32 blockHash = bytes32(1e10 + blockId);
+            bytes32 signalRoot = bytes32(1e9 + blockId);
             proveBlock(
                 Bob,
                 Bob,
@@ -524,10 +450,7 @@ contract TaikoL1LibTokenomicsTestnet is TaikoL1TestBase {
                 signalRoot
             );
             uint64 provenAt = uint64(block.timestamp);
-            console2.log(
-                "Proof reward is:",
-                L1.getProofReward(provenAt, proposedAt)
-            );
+            console2.log("Proof reward is:", L1.getProofReward(provenAt, proposedAt));
 
             verifyBlock(Carol, 1);
 
@@ -546,14 +469,13 @@ contract TaikoL1LibTokenomicsTestnet is TaikoL1TestBase {
     }
 
     /// @dev Test and see what happens when proof time is stable above the target and proving non consecutive
-    function test_balanced_state_reward_and_fee_if_proof_time_stable_above_target_proving_non_consecutive()
-        external
-    {
+    function test_balanced_state_reward_and_fee_if_proof_time_stable_above_target_proving_non_consecutive(
+    ) external {
         mine(1);
 
-        depositTaikoToken(Alice, 1E6 * 1E8, 100 ether);
-        depositTaikoToken(Bob, 1E6 * 1E8, 100 ether);
-        depositTaikoToken(Carol, 1E6 * 1E8, 100 ether);
+        depositTaikoToken(Alice, 1e6 * 1e8, 100 ether);
+        depositTaikoToken(Bob, 1e6 * 1e8, 100 ether);
+        depositTaikoToken(Carol, 1e6 * 1e8, 100 ether);
 
         TaikoData.BlockMetadata[] memory meta = new TaikoData.BlockMetadata[](
             30
@@ -582,8 +504,8 @@ contract TaikoL1LibTokenomicsTestnet is TaikoL1TestBase {
         mine(6);
         // Prove and verify
         for (uint256 blockId = 1; blockId < 30; blockId++) {
-            bytes32 blockHash = bytes32(1E10 + blockId);
-            bytes32 signalRoot = bytes32(1E9 + blockId);
+            bytes32 blockHash = bytes32(1e10 + blockId);
+            bytes32 signalRoot = bytes32(1e9 + blockId);
 
             proveBlock(
                 Bob,
@@ -597,10 +519,7 @@ contract TaikoL1LibTokenomicsTestnet is TaikoL1TestBase {
             );
 
             uint64 provenAt = uint64(block.timestamp);
-            console2.log(
-                "Proof reward is:",
-                L1.getProofReward(provenAt, proposedAt[blockId])
-            );
+            console2.log("Proof reward is:", L1.getProofReward(provenAt, proposedAt[blockId]));
 
             verifyBlock(Carol, 1);
             mine(5);
@@ -619,14 +538,13 @@ contract TaikoL1LibTokenomicsTestnet is TaikoL1TestBase {
     }
 
     /// @dev Test what happens when proof time decreases
-    function test_balanced_state_reward_and_fee_if_proof_time_decreasses_then_stabilizes_consecutive()
-        external
-    {
+    function test_balanced_state_reward_and_fee_if_proof_time_decreasses_then_stabilizes_consecutive(
+    ) external {
         mine(1);
 
-        depositTaikoToken(Alice, 1E6 * 1E8, 100 ether);
-        depositTaikoToken(Bob, 1E6 * 1E8, 100 ether);
-        depositTaikoToken(Carol, 1E6 * 1E8, 100 ether);
+        depositTaikoToken(Alice, 1e6 * 1e8, 100 ether);
+        depositTaikoToken(Bob, 1e6 * 1e8, 100 ether);
+        depositTaikoToken(Carol, 1e6 * 1e8, 100 ether);
 
         bytes32 parentHash = GENESIS_BLOCK_HASH;
 
@@ -639,17 +557,13 @@ contract TaikoL1LibTokenomicsTestnet is TaikoL1TestBase {
 
         //parentHash = prove_with_increasing_time(parentHash, 10);
         for (uint256 blockId = 1; blockId < 10; blockId++) {
-            TaikoData.BlockMetadata memory meta = proposeBlock(
-                Alice,
-                1000000,
-                1024
-            );
+            TaikoData.BlockMetadata memory meta = proposeBlock(Alice, 1000000, 1024);
             printVariables("after propose");
             uint64 proposedAt = uint64(block.timestamp);
             mine(11 - blockId);
 
-            bytes32 blockHash = bytes32(1E10 + blockId);
-            bytes32 signalRoot = bytes32(1E9 + blockId);
+            bytes32 blockHash = bytes32(1e10 + blockId);
+            bytes32 signalRoot = bytes32(1e9 + blockId);
             proveBlock(
                 Bob,
                 Bob,
@@ -661,10 +575,7 @@ contract TaikoL1LibTokenomicsTestnet is TaikoL1TestBase {
                 signalRoot
             );
             uint64 provenAt = uint64(block.timestamp);
-            console2.log(
-                "Proof reward is:",
-                L1.getProofReward(provenAt, proposedAt)
-            );
+            console2.log("Proof reward is:", L1.getProofReward(provenAt, proposedAt));
 
             verifyBlock(Carol, 1);
 
@@ -684,32 +595,16 @@ contract TaikoL1LibTokenomicsTestnet is TaikoL1TestBase {
         // Run another session with huge times
         for (uint256 blockId = 1; blockId < 10; blockId++) {
             printVariables("before propose");
-            TaikoData.BlockMetadata memory meta = proposeBlock(
-                Alice,
-                1000000,
-                1024
-            );
+            TaikoData.BlockMetadata memory meta = proposeBlock(Alice, 1000000, 1024);
             uint64 proposedAt = uint64(block.timestamp);
             printVariables("after propose");
             mine_proofTime();
 
-            bytes32 blockHash = bytes32(1E10 + blockId);
-            bytes32 signalRoot = bytes32(1E9 + blockId);
-            proveBlock(
-                Bob,
-                Bob,
-                meta,
-                parentHash,
-                1000000,
-                1000000,
-                blockHash,
-                signalRoot
-            );
+            bytes32 blockHash = bytes32(1e10 + blockId);
+            bytes32 signalRoot = bytes32(1e9 + blockId);
+            proveBlock(Bob, Bob, meta, parentHash, 1000000, 1000000, blockHash, signalRoot);
             uint64 provenAt = uint64(block.timestamp);
-            console2.log(
-                "Proof reward is:",
-                L1.getProofReward(provenAt, proposedAt)
-            );
+            console2.log("Proof reward is:", L1.getProofReward(provenAt, proposedAt));
             verifyBlock(Carol, 1);
 
             parentHash = blockHash;
@@ -723,17 +618,15 @@ contract TaikoL1LibTokenomicsTestnet is TaikoL1TestBase {
     }
 
     /// @dev Test what happens when proof time decreases
-    function test_balanced_state_reward_and_fee_if_proof_time_decreases_then_stabilizes_non_consecutive()
-        external
-    {
+    function test_balanced_state_reward_and_fee_if_proof_time_decreases_then_stabilizes_non_consecutive(
+    ) external {
         mine(1);
         // Requires a bit more tokens
-        depositTaikoToken(Alice, 1E8 * 1E8, 100 ether);
-        depositTaikoToken(Bob, 1E8 * 1E8, 100 ether);
-        depositTaikoToken(Carol, 1E8 * 1E8, 100 ether);
+        depositTaikoToken(Alice, 1e8 * 1e8, 100 ether);
+        depositTaikoToken(Bob, 1e8 * 1e8, 100 ether);
+        depositTaikoToken(Carol, 1e8 * 1e8, 100 ether);
 
-        TaikoData.BlockMetadata[]
-            memory metaArr = new TaikoData.BlockMetadata[](20);
+        TaikoData.BlockMetadata[] memory metaArr = new TaikoData.BlockMetadata[](20);
         uint64[] memory proposedAtArr = new uint64[](20);
 
         bytes32 parentHash = GENESIS_BLOCK_HASH;
@@ -757,8 +650,8 @@ contract TaikoL1LibTokenomicsTestnet is TaikoL1TestBase {
         mine(6);
         // Prove and verify
         for (uint256 blockId = 1; blockId < 10; blockId++) {
-            bytes32 blockHash = bytes32(1E10 + blockId);
-            bytes32 signalRoot = bytes32(1E9 + blockId);
+            bytes32 blockHash = bytes32(1e10 + blockId);
+            bytes32 signalRoot = bytes32(1e9 + blockId);
             proveBlock(
                 Bob,
                 Bob,
@@ -771,10 +664,7 @@ contract TaikoL1LibTokenomicsTestnet is TaikoL1TestBase {
             );
 
             uint64 provenAt = uint64(block.timestamp);
-            console2.log(
-                "Proof reward is:",
-                L1.getProofReward(provenAt, proposedAtArr[blockId])
-            );
+            console2.log("Proof reward is:", L1.getProofReward(provenAt, proposedAtArr[blockId]));
 
             verifyBlock(Carol, 1);
             mine(21 - blockId);
@@ -793,32 +683,16 @@ contract TaikoL1LibTokenomicsTestnet is TaikoL1TestBase {
         // Run another session with huge times
         for (uint256 blockId = 1; blockId < 10; blockId++) {
             printVariables("before propose");
-            TaikoData.BlockMetadata memory meta = proposeBlock(
-                Alice,
-                1000000,
-                1024
-            );
+            TaikoData.BlockMetadata memory meta = proposeBlock(Alice, 1000000, 1024);
             uint64 proposedAt = uint64(block.timestamp);
             printVariables("after propose");
             mine_proofTime();
 
-            bytes32 blockHash = bytes32(1E10 + blockId);
-            bytes32 signalRoot = bytes32(1E9 + blockId);
-            proveBlock(
-                Bob,
-                Bob,
-                meta,
-                parentHash,
-                1000000,
-                1000000,
-                blockHash,
-                signalRoot
-            );
+            bytes32 blockHash = bytes32(1e10 + blockId);
+            bytes32 signalRoot = bytes32(1e9 + blockId);
+            proveBlock(Bob, Bob, meta, parentHash, 1000000, 1000000, blockHash, signalRoot);
             uint64 provenAt = uint64(block.timestamp);
-            console2.log(
-                "Proof reward is:",
-                L1.getProofReward(provenAt, proposedAt)
-            );
+            console2.log("Proof reward is:", L1.getProofReward(provenAt, proposedAt));
             verifyBlock(Carol, 1);
 
             parentHash = blockHash;
@@ -834,14 +708,12 @@ contract TaikoL1LibTokenomicsTestnet is TaikoL1TestBase {
     }
 
     /// @dev Test and see what happens when proof time is stable at the target and proving consecutive
-    function test_balanced_state_reward_and_fee_if_proof_time_stable_consecutive()
-        external
-    {
+    function test_balanced_state_reward_and_fee_if_proof_time_stable_consecutive() external {
         mine(1);
 
-        depositTaikoToken(Alice, 1E6 * 1E8, 100 ether);
-        depositTaikoToken(Bob, 1E6 * 1E8, 100 ether);
-        depositTaikoToken(Carol, 1E6 * 1E8, 100 ether);
+        depositTaikoToken(Alice, 1e6 * 1e8, 100 ether);
+        depositTaikoToken(Bob, 1e6 * 1e8, 100 ether);
+        depositTaikoToken(Carol, 1e6 * 1e8, 100 ether);
 
         bytes32 parentHash = GENESIS_BLOCK_HASH;
 
@@ -855,17 +727,13 @@ contract TaikoL1LibTokenomicsTestnet is TaikoL1TestBase {
         //parentHash = prove_with_increasing_time(parentHash, 10);
         for (uint256 blockId = 1; blockId < 10; blockId++) {
             printVariables("before propose");
-            TaikoData.BlockMetadata memory meta = proposeBlock(
-                Alice,
-                1000000,
-                1024
-            );
+            TaikoData.BlockMetadata memory meta = proposeBlock(Alice, 1000000, 1024);
             uint64 proposedAt = uint64(block.timestamp);
             printVariables("after propose");
             mine_proofTime();
 
-            bytes32 blockHash = bytes32(1E10 + blockId);
-            bytes32 signalRoot = bytes32(1E9 + blockId);
+            bytes32 blockHash = bytes32(1e10 + blockId);
+            bytes32 signalRoot = bytes32(1e9 + blockId);
             proveBlock(
                 Bob,
                 Bob,
@@ -877,10 +745,7 @@ contract TaikoL1LibTokenomicsTestnet is TaikoL1TestBase {
                 signalRoot
             );
             uint64 provenAt = uint64(block.timestamp);
-            console2.log(
-                "Proof reward is:",
-                L1.getProofReward(provenAt, proposedAt)
-            );
+            console2.log("Proof reward is:", L1.getProofReward(provenAt, proposedAt));
 
             verifyBlock(Carol, 1);
 
@@ -899,9 +764,7 @@ contract TaikoL1LibTokenomicsTestnet is TaikoL1TestBase {
     }
 
     /// @dev Test a scenario which very close to a testnet behaviour
-    function test_balanced_state_reward_and_fee_if_proof_time_stable_non_consecutive()
-        external
-    {
+    function test_balanced_state_reward_and_fee_if_proof_time_stable_non_consecutive() external {
         uint256 Alice_start_balance = L1.getTaikoTokenBalance(Alice);
         uint256 Bob_start_balance = L1.getTaikoTokenBalance(Bob);
 
@@ -925,8 +788,8 @@ contract TaikoL1LibTokenomicsTestnet is TaikoL1TestBase {
                 proposedAt[blockId] = (uint64(block.timestamp));
                 printVariables("after propose");
 
-                blockHashes[blockId] = bytes32(1E10 + blockId); //blockHash;
-                signalRoots[blockId] = bytes32(1E9 + blockId); //signalRoot;
+                blockHashes[blockId] = bytes32(1e10 + blockId); //blockHash;
+                signalRoots[blockId] = bytes32(1e9 + blockId); //signalRoot;
 
                 if (blockId > proofTime) {
                     //Start proving with an offset
@@ -944,10 +807,7 @@ contract TaikoL1LibTokenomicsTestnet is TaikoL1TestBase {
                     uint64 provenAt = uint64(block.timestamp);
                     console2.log(
                         "Proof reward is:",
-                        L1.getProofReward(
-                            provenAt,
-                            proposedAt[blockId - proofTime]
-                        )
+                        L1.getProofReward(provenAt, proposedAt[blockId - proofTime])
                     );
                 }